Mouse prende l'oggetto e lo deposito sulla finestra

+++ src/main/java/jade/Prefabs.java
@@ -6,7 +6,7 @@ import org.joml.Vector2f;
 
 public class Prefabs {
 
-    public static GameObject generateSpriteObject(Sprite sprite, int sizeX, int sizeY) {
+    public static GameObject generateSpriteObject(Sprite sprite, float sizeX, float sizeY) {
         GameObject block = new GameObject("Sprite_Object_Gen"+sprite.getTexId(),
                 new Transform(new Vector2f(),new Vector2f(sizeX,sizeY)),0);
         //TRANFORM PIAZZATO A 0,0 CON LE SIZE DELLO SPRITE COME SCALA
